+++ peakrack_kyc/lang/chinese-hk.php
@@ -1,0 +1,16 @@
+<?php
+// SPDX-License-Identifier: LicenseRef-PeakRack-Proprietary
+
+/**
+ * PeakRack KYC for WHMCS
+ *
+ * Official repository:
+ * https://github.com/Techshrr/whmcs_peakrack_kyc
+ *
+ * Copyright (c) 2026 PeakRack. All rights reserved.
+ * Unauthorized copying, modification, distribution, sublicensing, or commercial use
+ * is prohibited without prior written permission.
+ */
+
+$_ADDONLANG['module_name'] = 'PeakRack 實名認證';
+$_ADDONLANG['module_description'] = '適用於 WHMCS 的實名認證、證件上傳、人工審核和指定產品強制實名模組。';